Provide standalone server + Docker
Problem: We are running tests on multiple shards (up to 12) and want to try globalCache to speed up the pipeline. The proposed setup/teardown option does not suit us, as we want the data to be shared between all instances.

Solution: run the cache as a separate server during test execution and “kill” it after the run.